packages/rag_engine/requirements.txt
# RAG Engine Dependencies
# PDF extraction
pymupdf>=1.24.0

# Embeddings
openai>=1.0.0
tiktoken>=0.5.0

# Vector store
chromadb>=0.4.0

# Keyword search
rank-bm25>=0.2.2

# Data validation
pydantic>=2.5.0

# Configuration
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# CLI
click>=8.1.0

# Async utilities
aiofiles>=23.2.0

# Numerical operations
numpy>=1.24.0
